+++ common/lib/util_spi.h
@@ -28,7 +28,7 @@ uint8_t fw_update(uint32_t offset, uint16_t msg_len, uint8_t *msg_buf, bool sect
 uint8_t fw_update_cxl(uint8_t flash_position);
 
 int pal_get_bios_flash_position();
-bool pal_switch_bios_spi_mux(int gpio_status);
+bool pal_switch_bios_spi_mux(uint8_t gpio_status);
 int pal_get_cxl_flash_position();
 bool pal_switch_cxl_spi_mux();
